+++ src/emsarray/cli/__init__.py
@@ -10,18 +10,20 @@ import emsarray
 from ._operation import Operation
 from .clip import Clip
 from .exceptions import CommandException
+from .extract_points import ExtractPoints
 from .utils import console_entrypoint
 
 __all__ = ['main', 'CommandException', 'Operation']
 
 
 def command_line_flags(parser: argparse.ArgumentParser) -> None:
-    parser.set_defaults(func=lambda: print(parser.format_help()))
+    parser.set_defaults(func=lambda x: print(parser.format_help()))
     parser.add_argument(
         '-V', '--version', action='version', version=f'%(prog)s {emsarray.__version__}')
 
     subparsers = parser.add_subparsers(title="Operations")
     Clip().add_parser(subparsers)
+    ExtractPoints().add_parser(subparsers)
 
 
 @console_entrypoint(command_line_flags)